Reduce bucket relays around UserList
